+++ simulation/executor/simulate_dev.go
@@ -69,7 +69,8 @@ func newSimulatorState(simParams Params, initialHeader tmproto.Header, tb testin
 func (simState *simState) SimulateAllBlocks(
 	w io.Writer,
 	simCtx *simtypes.SimCtx,
-	blockSimulator blockSimFn) (stopEarly bool, err error) {
+	blockSimulator blockSimFn,
+) (stopEarly bool, err error) {
 	stopEarly = false
 	initialHeight := simState.config.InitializationConfig.InitialBlockHeight
 	numBlocks := simState.config.NumBlocks
@@ -101,7 +102,7 @@ func (simState *simState) SimulateBlock(simCtx *simtypes.SimCtx, blockSimulator 
 	}
 
 	requestBeginBlock := simState.beginBlock(simCtx)
-	ctx := simCtx.BaseApp().NewContext(false, simState.header)
+	ctx := simCtx.BaseApp().NewContext(false, simState.header).WithBlockTime(simState.header.Time)
 
 	// Run queued operations. Ignores blocksize if blocksize is too small
 	numQueuedOpsRan, err := simState.runQueuedOperations(simCtx, ctx)